refactor(dev_pipeline): extract dev_pipeline_lib + fold in #1392 Fix 2 (false-green guard)
STEP 0 of the dev-pipeline reconciler. A pure, behavior-preserving extraction
of the proven dev-pipeline helpers out of `dev_pipeline.py`'s monolithic `_BODY`
raw-string into a new shared source-string module `dev_pipeline_lib.py`
(`DEV_PIPELINE_LIB`), so BOTH the monolith and the coming reconciler import one
source of truth — the same authoring-time source-string sharing pattern as
`gh_body.GH_BODY_HELPERS` / `comment_idempotency.COMMENT_IDEMPOTENCY_HELPERS`.

Extracted (byte-identical): the scripted spec gate (`spec_ok` + `_strip_meta`/
`_comment_texts`/`_marker_resolved`/`_find_blocker`), the gh/pagination plumbing
incl. the #1156 full-thread read (`gh`/`_gh_once`/`_is_transient`/`_headers`/
`_link_next_page`/`_with_query`/`_status`/`_ipath`/`post_markered_comment`), the
SHA/commit-identity checks (`_pr_head_sha`/`_is_sha1`/`_resolve_sha1`), the
fail-closed merge guard + rebase (`_merge_guard_command`/`_ruff_format_parity_
sentinels`/`_needs_rebase`/`_rebase_command`/`_mark_ready_query`), the
deterministic risk floor (`_risk_floor`/`_changed_workflow_files`/
`_is_workflow_path`), `_find_open_pr`, the label/terminal/close-on-merge plumbing
(`_label`/`_unlabel`/`_fail`/`_close_source_issue`/`_file_followup`), and the
bounded CI-watch + sync/rebase loops (`_watch_ci`/`_live_head_sha`/`_sync_branch`),
plus `_unwrap`/`_owner`. Only `main()` (the state machine) stays in `_BODY`.

Fold in aios#1392's Fix 2 ONLY (the false-green guard): new `_ci_verdict`
(3-state pass/red/retry — a green/no_ci is trusted only when the polled commit IS
the live PR head (#1314) AND the required-check set concluded (#1364), fail-closed
to retry when neither anchor is available), new `_shas_equal`, and two OPTIONAL
`CI_SCHEMA` fields (`polled_sha`/`required_complete`, both fail-closed when
omitted). `_watch_ci` routes the agent's green/no_ci through `_ci_verdict` (pass →
proceed, retry → re-watch without dispatching a fixer, red → fix) while KEEPING its
original `(ci_ok, ci_agent_error, head_sha)` signature and `main()`'s verify-gate
escalation — so no #1392 Fix-1 surface is introduced.

#1392's Fix 1 (the run_completion[errored] recovery / `autodev:recovery-N` label
budget / `_recover_or_park`/`_recovery_attempts`/`_handle_ci_outcome`) is
deliberately NOT carried — superseded by the reconciler design. Once this lands,
#1392 can be closed (Fix 2 lives here; Fix 1 is dropped).

Proven behavior-preserving by AST symbol-diff of the assembled script: only
`_ci_verdict`/`_shas_equal` added, only `CI_SCHEMA`/`_watch_ci` changed; the other
87 symbols and `main()` are byte-identical. The full existing dev_pipeline suite
(137 tests across the fixture + rebase/close/retry/risk-floor/ruff/surface units)
passes unchanged — the only test-harness change is a `Scenario._ci_verdict` helper
modelling a correct watch agent (defaults polled_sha + required_complete) so the
simulated green is verifiable. 13 new focused unit tests cover `_shas_equal` (5)
and `_ci_verdict` (8); the in-envelope `retry_count`/`MAX_RUN_RETRIES` dead-letter
guard (pre-existing, NOT Fix 1) is retained unchanged.

DO NOT auto-merge — the seat reviews adversarially, then deploys (re-registers the
dev-pipeline + triage workflows so the shared lib reaches the live worker).
